refactor(PrimaryButton): split up the file into a own folder
